pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Santosh G",email = "santosh.g00028@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.124.4,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "supabase (>=2.25.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "celery (>=5.6.0,<6.0.0)",
    "clerk-backend-api (>=4.2.0,<5.0.0)",
    "boto3 (>=1.42.9,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
